PawPal App/requirements.txt
streamlit>=1.33,<2
pandas>=2.0,<3
pytest>=7.0,<9
openai>=1.40,<2
python-dotenv>=1.0,<2
numpy>=1.26,<3
scikit-learn>=1.4,<2
